kodi.packages.buildKodiBinaryAddon: add extraCMakeFlags arg, add symlinks for libretro cores

+++ pkgs/applications/video/kodi/build-kodi-binary-addon.nix
@@ -5,6 +5,7 @@
 , extraNativeBuildInputs ? []
 , extraBuildInputs ? []
 , extraRuntimeDependencies ? []
+, extraCMakeFlags ? []
 , extraInstallPhase ? "", ... } @ attrs:
 toKodiAddon (stdenv.mkDerivation ({
   name = "kodi-" + name;
@@ -19,7 +20,7 @@ toKodiAddon (stdenv.mkDerivation ({
   # disables check ensuring install prefix is that of kodi
   cmakeFlags = [
     "-DOVERRIDE_PATHS=1"
-  ];
+  ] ++ extraCMakeFlags;
 
   # kodi checks for addon .so libs existance in the addon folder (share/...)
   # and the non-wrapped kodi lib/... folder before even trying to dlopen
@@ -28,7 +29,10 @@ toKodiAddon (stdenv.mkDerivation ({
     runHook preInstall
 
     make install
-    ln -s $out/lib/addons/${n}/${n}.so.${version} $out${addonDir}/${n}/${n}.so.${version}
+
+    [[ -f $out/lib/addons/${n}/${n}.so ]] && ln -s $out/lib/addons/${n}/${n}.so $out${addonDir}/${n}/${n}.so || true
+    [[ -f $out/lib/addons/${n}/${n}.so.${version} ]] && ln -s $out/lib/addons/${n}/${n}.so.${version} $out${addonDir}/${n}/${n}.so.${version} || true
+
     ${extraInstallPhase}
 
     runHook postInstall
